Move 'BetterConfigParser' class into a separate file. Removed unused 'pprint' import.
